+++ SAST/codeql/codeql_v2_9_2/Dockerfile
@@ -10,7 +10,6 @@ ARG CODEQL_INTERFACE_DIR="${TPF_HOME}/SAST/codeql/codeql_v2_9_2"
 COPY ./__init__.py ${CODEQL_INTERFACE_DIR}/__init__.py
 COPY ./config.yaml ${CODEQL_INTERFACE_DIR}/config.yaml
 COPY ./codeql.py ${CODEQL_INTERFACE_DIR}/codeql.py
-COPY ./resources ${CODEQL_INTERFACE_DIR}/resources
 
 RUN apt-get update
 RUN apt-get install wget
