--- SAST/codeql/codeql_v2_9_2/resources/_template_build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-find $PATTERNSRC -name "*.java" > $TPFOUT/sources.txt
-javac -classpath $PATTERNLIB/*.jar  @$TPFOUT/sources.txt
